// src/api/generated/models/AccountSearchResultDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AccountSearchResultDto = {
    /**
     * The matched account address.
     */
    address: string;
    /**
     * The account's `.chain` name, if any.
     */
    chain_name: string | null;
};